+++ apps/backend/internal/domain/keys/platform_key_create.go
@@ -48,7 +48,7 @@ func (s *service) CreatePlatformKey(ctx context.Context, input types.CreatePlatf
 		if msg := common.ValidateModelIDsForMember(*input.MemberID, input.ModelWhitelist, members, departments, rules, models, common.ModelNotInDeptMessage); msg != nil {
 			return types.PlatformKey{}, domain.Validation(*msg)
 		}
-		if msg := budget.ValidateMemberScopeKeyBudget(members, platformKeys, *input.MemberID, input.Budget, uuid.Nil); msg != nil {
+		if msg := budget.ValidateMemberScopeKeyBudget(members, platformKeys, *input.MemberID, int64(input.Budget), uuid.Nil); msg != nil {
 			return types.PlatformKey{}, domain.Validation(*msg)
 		}
 	case types.PlatformKeyScopeProject:
@@ -56,7 +56,7 @@ func (s *service) CreatePlatformKey(ctx context.Context, input types.CreatePlatf
 		if !ok {
 			return types.PlatformKey{}, domain.NotFound("Project not found")
 		}
-		if msg := budget.ValidateProjectScopeKeyBudget(input.Scope, project, platformKeys, input.MemberID, input.Budget, uuid.Nil); msg != nil {
+		if msg := budget.ValidateProjectScopeKeyBudget(input.Scope, project, platformKeys, input.MemberID, int64(input.Budget), uuid.Nil); msg != nil {
 			return types.PlatformKey{}, domain.Validation(*msg)
 		}
 		if input.MemberID != nil {
@@ -75,7 +75,7 @@ func (s *service) CreatePlatformKey(ctx context.Context, input types.CreatePlatf
 		if msg := common.ValidateModelIDsForMember(*input.MemberID, input.ModelWhitelist, members, departments, rules, models, common.ModelNotInDeptMessage); msg != nil {
 			return types.PlatformKey{}, domain.Validation(*msg)
 		}
-		if msg := budget.ValidateProjectScopeKeyBudget(input.Scope, project, platformKeys, input.MemberID, input.Budget, uuid.Nil); msg != nil {
+		if msg := budget.ValidateProjectScopeKeyBudget(input.Scope, project, platformKeys, input.MemberID, int64(input.Budget), uuid.Nil); msg != nil {
 			return types.PlatformKey{}, domain.Validation(*msg)
 		}
 	}
@@ -88,7 +88,7 @@ func (s *service) CreatePlatformKey(ctx context.Context, input types.CreatePlatf
 		ID:   uuid.Must(uuid.NewV7()),
 		Name: input.Name, KeyPrefix: "pending...", Scope: input.Scope,
 		MemberID: input.MemberID, ProjectID: input.ProjectID,
-		Status: "active", Budget: input.Budget, Consumed: 0,
+		Status: "active", Budget: int64(input.Budget), Consumed: 0,
 		ModelWhitelist: append([]uuid.UUID{}, input.ModelWhitelist...),
 		CreatedAt:      time.Now().Format("2006-01-02"),
 	}
