+++ apps/backend/internal/pkg/budget/lookup.go
@@ -5,7 +5,7 @@ import (
 	"github.com/tokenjoy/backend/internal/domain/types"
 )
 
-func GetReservedPoolForDepartment(tree []types.BudgetNode, departmentID uuid.UUID) float64 {
+func GetReservedPoolForDepartment(tree []types.BudgetNode, departmentID uuid.UUID) int64 {
 	node := FindBudgetNode(tree, departmentID)
 	if node == nil || node.ReservedPool == nil {
 		return 0
@@ -13,7 +13,7 @@ func GetReservedPoolForDepartment(tree []types.BudgetNode, departmentID uuid.UUI
 	return *node.ReservedPool
 }
 
-func GetReservedPoolForMember(tree []types.BudgetNode, members []types.Member, memberID uuid.UUID) float64 {
+func GetReservedPoolForMember(tree []types.BudgetNode, members []types.Member, memberID uuid.UUID) int64 {
 	for _, member := range members {
 		if member.ID == memberID {
 			return GetReservedPoolForDepartment(tree, member.DepartmentID)
