+++ apps/backend/internal/store/budget_repo.go
@@ -11,8 +11,8 @@ import (
 type BudgetRepository interface {
 	AcquireBudgetLock(ctx context.Context) error
 	OrgNodeBudget() OrgNodeBudgetRepository
-	GetProjectBudget(ctx context.Context, projectID uuid.UUID) (budget, consumed float64, found bool, err error)
-	GetProjectMemberBudget(ctx context.Context, projectID, memberID uuid.UUID) (float64, bool, error)
+	GetProjectBudget(ctx context.Context, projectID uuid.UUID) (budget, consumed int64, found bool, err error)
+	GetProjectMemberBudget(ctx context.Context, projectID, memberID uuid.UUID) (int64, bool, error)
 	Projects(ctx context.Context) ([]types.Project, error)
 	SetProjects(ctx context.Context, projects []types.Project) error
 	OverrunPolicy(ctx context.Context) (types.OverrunPolicyConfig, error)
