+++ apps/backend/internal/store/postgres/usage_aggregate.go
@@ -77,7 +77,7 @@ func aggregateUsageRows(
 		existing.DepartmentID = key.DepartmentID
 		existing.MemberID = key.MemberID
 		existing.Model = key.Model
-		existing.Cost += row.Cost
+		existing.Cost += float64(row.Cost)
 		existing.DisplayCost += row.DisplayCost
 		existing.CallCount += row.CallCount
 		existing.InputTokens += row.InputTokens
@@ -125,7 +125,7 @@ func summaryUsageTotals(rows []types.UsageBucketRow, start, end time.Time) types
 		if row.BucketStart.Before(start) || !row.BucketStart.Before(end) {
 			continue
 		}
-		totals.Cost += row.Cost
+		totals.Cost += float64(row.Cost)
 		totals.DisplayCost += row.DisplayCost
 		totals.CallCount += row.CallCount
 		totals.InputTokens += row.InputTokens
